+++ src/HealResult.ts
@@ -1,0 +1,5 @@
+import { Character } from "./character/Character";
+
+export type HealResult =
+  | { outcome: "healed"; amount: number; target: Character }
+  | { outcome: "cooldown" };
