+++ src/character/isCharacterOnCooldown.ts
@@ -4,4 +4,6 @@ import { getCooldownRemaining } from "../getCooldownRemaining";
 export const isCharacterOnCooldown = (
   characterId: string,
   type: keyof Character["cooldowns"]
-): boolean => (getCooldownRemaining(characterId, type) ?? 0) > 0;
+): boolean => {
+  return (getCooldownRemaining(characterId, type) ?? 0) > 0;
+};
