+++ src/commands/encounters/tavern/barFight.ts
@@ -14,8 +14,8 @@ export async function barFight(interaction: CommandInteraction): Promise<void> {
     .setTitle("Bar Fight!")
     .setColor("RED")
     .setDescription("You get into a drunken brawl and are kicked out.")
-    .addField("HP Lost", damage.toString())
-    .addField("XP Gained", "1")
+    .addField("HP Lost", "🩸" + damage.toString())
+    .addField("XP Gained", ":xp:1")
     .setImage("https://i.imgur.com/yo1JymD.png");
   const character = getUserCharacter(interaction.user);
   if (character.hp > 0 && character.quests.survivor) {
