+++ src/isStatusEffectExpired.ts
@@ -1,0 +1,4 @@
+import { StatusEffect } from "./statusEffects/StatusEffect";
+
+export const isStatusEffectExpired = (effect: StatusEffect): boolean =>
+  Date.now() > new Date(effect.started).valueOf() + effect.duration;
